refactor modal to be a better experience
