+++ components/Shared/ModalContent/HeightByPeriodHelp/index.tsx
@@ -1,0 +1,12 @@
+import styled from "@emotion/styled";
+import React from "react";
+
+const BuoyHelpModalContent: React.FC = () => {
+  return (
+    <>
+      <p>Explain what the multicomponent sea state is!</p>
+    </>
+  );
+};
+
+export default BuoyHelpModalContent;
